+++ src/plugins/saved_objects_management/public/lib/fetch_export_objects.test.ts
@@ -1,0 +1,59 @@
+/*
+ * Copyright OpenSearch Contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+import { fetchExportObjects } from './fetch_export_objects';
+import { httpServiceMock } from '../../../../core/public/mocks';
+
+describe('fetchExportObjects', () => {
+  it('make http call with body provided', async () => {
+    const httpClient = httpServiceMock.createStartContract();
+    await fetchExportObjects(httpClient, [], false, {
+      workspaces: ['foo'],
+    });
+    expect(httpClient.post).toMatchInlineSnapshot(`
+      [MockFunction] {
+        "calls": Array [
+          Array [
+            "/api/saved_objects/_export",
+            Object {
+              "body": "{\\"workspaces\\":[\\"foo\\"],\\"objects\\":[],\\"includeReferencesDeep\\":false}",
+            },
+          ],
+        ],
+        "results": Array [
+          Object {
+            "type": "return",
+            "value": undefined,
+          },
+        ],
+      }
+    `);
+  });
+
+  it('make http call with body provided and includeReferencesDeep is undefined', async () => {
+    const httpClient = httpServiceMock.createStartContract();
+    await fetchExportObjects(httpClient, [], undefined, {
+      workspaces: ['foo'],
+    });
+    expect(httpClient.post).toMatchInlineSnapshot(`
+      [MockFunction] {
+        "calls": Array [
+          Array [
+            "/api/saved_objects/_export",
+            Object {
+              "body": "{\\"workspaces\\":[\\"foo\\"],\\"objects\\":[],\\"includeReferencesDeep\\":false}",
+            },
+          ],
+        ],
+        "results": Array [
+          Object {
+            "type": "return",
+            "value": undefined,
+          },
+        ],
+      }
+    `);
+  });
+});